+++ CHANGELOG.md
@@ -70,7 +70,7 @@ Release
   - ```frame_data``` can only contain ```Num``` instances
   - ```object_data``` can not contain ```Num``` instances anymore
 - Major restructuring of the project directories
-- ```FrameInterval.from_frame_uids()```: create ```FrameIntervals``` by providing a list of frame uids
+- ```FrameInterval.from_frame_ids()```: create ```FrameIntervals``` by providing a list of frame uids
 - ```Object.object_data_pointers()```: generate ```ElementDataPointers```
 - ```Scene.frame_intervals()```, ```Object.frame_intervals()```: generate ```FrameIntervals```
 - ```Object.asdict()``` now provides also frame intervals and object data pointers, if the frames from the scene are provided